gpu: pvr: Add thread safety checks to SGX MMU code.
The SGX MMU code was not designed to be used from multiple driver
threads. One theory is that somehow the driver does sometimes use
the MMU code from multiple threads, which could cause corruption
of meta data.

We must allow PID 1 through because the kernel uses this ID before
userspace is up, and the module_init code interacts with the MMU
subsystem.

Otherwise, check that the outer bridge lock is held. If it is held,
check the PID of the caller matches the PID holding the bridge lock.

Change-Id: I35f42f251316b158576e5d853f8099e24942b4f2